--- vm/cli/commands/clean.rs
@@ -1,36 +1,0 @@
-// Copyright (c) 2019-2025 Provable Inc.
-// This file is part of the snarkVM library.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at:
-
-// http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use super::*;
-
-/// Cleans the Aleo package build directory.
-#[derive(Debug, Parser)]
-pub struct Clean;
-
-impl Clean {
-    /// Cleans an Aleo package build directory.
-    pub fn parse(self) -> Result<String> {
-        // Derive the program directory path.
-        let path = std::env::current_dir()?;
-
-        // Clean the build directory.
-        Package::<CurrentNetwork>::clean(&path)?;
-
-        // Prepare the path string.
-        let path_string = format!("(in \"{}\")", path.join("build").display());
-
-        Ok(format!("✅ Cleaned the build directory {}", path_string.dimmed()))
-    }
-}
