--- vm/cli/errors.rs
@@ -1,29 +1,0 @@
-// Copyright (c) 2019-2025 Provable Inc.
-// This file is part of the snarkVM library.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at:
-
-// http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#[derive(Debug, Error)]
-pub enum UpdaterError {
-    #[error("{}: {}", _0, _1)]
-    Crate(&'static str, String),
-
-    #[error("The current version {} is more recent than the release version {}", _0, _1)]
-    OldReleaseVersion(String, String),
-}
-
-impl From<self_update::errors::Error> for UpdaterError {
-    fn from(error: self_update::errors::Error) -> Self {
-        UpdaterError::Crate("self_update", error.to_string())
-    }
-}
