--- vm/cli/mod.rs
@@ -1,26 +1,0 @@
-// Copyright (c) 2019-2025 Provable Inc.
-// This file is part of the snarkVM library.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at:
-
-// http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-mod cli;
-pub use cli::*;
-
-mod commands;
-pub use commands::*;
-
-mod errors;
-pub use errors::*;
-
-pub mod helpers;
-pub use helpers::*;
